+++ packages/orq-rc/RELEASES.md
@@ -11169,3 +11169,13 @@ Based on:
 - [python v4.16.0-rc.15] packages/orq-rc
 ### Releases
 - [PyPI v4.16.0-rc.15] https://pypi.org/project/orq-ai-sdk/4.16.0-rc.15 - packages/orq-rc
+
+## 2026-09-23 15:39:07
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.798.0 (2.938.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.16.0-rc.16] packages/orq-rc
+### Releases
+- [PyPI v4.16.0-rc.16] https://pypi.org/project/orq-ai-sdk/4.16.0-rc.16 - packages/orq-rc